refactor out writing tasks from segment
